+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Introduced Network Transaction Builder (#840).
 - Added way of executing and proving network transactions (#841).
 - Added options for configuring NTB in `bundled` command (#884).
+- Proof-of-Work difficulty is now adjusted based on the number of concurrent requests (#865).
 
 ## v0.8.2 (2025-05-04)